--- src/shared/GlobalTopBanners/AppInstallBanner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppInstallBanner'
